Fix for #6845
The current on-disk format for encrypted datasets protects
not only the encrypted and authenticated blocks, but also
the order and interpretation of these blocks. In order to
make this work while maintaining the ability to do raw sends
the indirect bps maintain a secure checksum of all the MACs
in the block below it, along with a few other fields that
determine how the data is interpretted.

Unfortunately, the current on-disk format erroniously
includes the byteorder and compression of the blocks below,
which is not portable and thus cannot support raw sends.
Unfortunately, it is also not possible to easily work around
this issue due to a separate and much smaller bug which
causes indirect blocks for dnodes to not be compressed.

This patch zero's out the byteorder and compression when
computing the MAC (as they should have been) and registers
an errata for the on-disk format bug.

Signed-off-by: Tom Caputi <tcaputi@datto.com>