Fix for raw send with hole_birth
In order to protect the structure of data alongside
the data itself in an encrypted dataset, the encyrption
feature uses a combination of cryptographic hashes and
message authentication codes to protect certain info
about the structure of data. However, we can only use
this technique to protect data that can be replicated
on another system when doing a raw 'zfs send'.

Unfortunately, many of the bits of bp->blk_prop which
are protected in this way may be treated differently
if they are holes that are sent to a pool that uses
the hole_birth feature, which breaks the authentication
mechanism. The fix is to explicitly zero these fields
for the purposes of the authentication code. This
change does represent a break in the on-disk format.

TODO: Add errata handling

Signed-off-by: Tom Caputi <tcaputi@datto.com>